Made SurfaceGlow a child of Surface because scaling issues
So I changed the celestial node to be scaled by distance, but not radius, and the surface node to be scaled by radius, but not distance. Then the SurfaceGlow needs to also be scaled by radius, so it needs to be a child of the surface node. Does that make sense? Oh and also I'm using this script to set static values of other classes that need a reference to a node to get a node in the scene; probably a bad idea, but I did it anyway.